requirements.txt
Django~=3.2
PyYAML==6.0.1
setuptools==68.2.2
wheel==0.41.2
twine==4.0.2
tox==4.11.3
